Added .vimrc file and added some aliases in zshrc
This is all to be used on the mac at work and on hyak. I will eventually make
a new branch for the work computer. Overall .vimrc is is much more organized
and .zshrc has many shortcuts to directories that are only available at work
so parts of this will need to be cherry picked once I move to a computer that
isn't at work